+++ sbroad/sbroad-core/src/executor.rs
@@ -55,7 +55,7 @@ impl Plan {
     /// # Errors
     /// - Failed to optimize the plan.
     pub fn optimize(&mut self) -> Result<(), SbroadError> {
-        self.update_local_timestamps()?;
+        self.update_timestamps()?;
         self.cast_constants()?;
         self.replace_in_operator()?;
         self.push_down_not()?;